refactoring, debugging rename-database.xq

+++ src/main/java/de/axxepta/oxygen/actions/DeleteAction.java
@@ -3,8 +3,7 @@ package de.axxepta.oxygen.actions;
 import de.axxepta.oxygen.api.BaseXConnectionWrapper;
 import de.axxepta.oxygen.api.BaseXSource;
 import de.axxepta.oxygen.api.Connection;
-import de.axxepta.oxygen.tree.BasexTree;
-import de.axxepta.oxygen.tree.TreeListener;
+import de.axxepta.oxygen.tree.ArgonTree;
 import de.axxepta.oxygen.tree.TreeUtils;
 
 import javax.swing.*;
@@ -19,67 +18,66 @@ import java.awt.event.ActionEvent;
  */
 public class DeleteAction extends AbstractAction {
 
-    private BasexTree tree;
+    private ArgonTree tree;
     private TreeModel treeModel;
-    private TreeListener treeListener;
 
-    public DeleteAction(String name, Icon icon, BasexTree tree, TreeListener treeListener){
+    public DeleteAction(String name, Icon icon, ArgonTree tree){
         super(name, icon);
         this.tree = tree;
         this.treeModel = tree.getModel();
-        this.treeListener = treeListener;
     }
 
-    public DeleteAction(BasexTree tree, TreeListener treeListener){
+    public DeleteAction(ArgonTree tree){
         super();
         this.tree = tree;
         this.treeModel = tree.getModel();
-        this.treeListener = treeListener;
     }
 
     @Override
     public void actionPerformed(ActionEvent e) {
         TreePath[] paths = tree.getSelectionPaths();
         boolean deleteAll = false;
-        for (TreePath path : paths) {
-            BaseXSource source = TreeUtils.sourceFromTreePath(path);
-            String db_path = TreeUtils.resourceFromTreePath(path);
-            if ((source != null) && (!db_path.equals(""))) {
-                // don't try to delete databases!
-                if ((!(source == BaseXSource.DATABASE)) || (db_path.contains("/"))) {
-                    if (((DefaultMutableTreeNode) path.getLastPathComponent()).getAllowsChildren()) {
-                        if (tree.isCollapsed(path)) {
-                            tree.expandPath(path);
-                            tree.collapsePath(path);
+        if (paths != null) {
+            for (TreePath path : paths) {
+                BaseXSource source = TreeUtils.sourceFromTreePath(path);
+                String db_path = TreeUtils.resourceFromTreePath(path);
+                if ((source != null) && (!db_path.equals(""))) {
+                    // don't try to delete databases!
+                    if ((!(source == BaseXSource.DATABASE)) || (db_path.contains("/"))) {
+                        if (((DefaultMutableTreeNode) path.getLastPathComponent()).getAllowsChildren()) {
+                            if (tree.isCollapsed(path)) {
+                                tree.expandPath(path);
+                                tree.collapsePath(path);
+                            } else {
+                                tree.collapsePath(path);
+                                tree.expandPath(path);
+                            }
+                        }
+                        if (treeModel.getChildCount(path.getLastPathComponent()) == 0) {
+
+                            int dialogResult = JOptionPane.YES_OPTION;
+                            if (!deleteAll) {
+                                Object[] answerOptions = {"Yes", "All", "No"};
+                                dialogResult = JOptionPane.showOptionDialog(null, "Do you really want to delete the file\n" +
+                                                TreeUtils.urlStringFromTreePath(path) + "?", "Delete Resource(s)",
+                                        JOptionPane.YES_NO_CANCEL_OPTION, JOptionPane.QUESTION_MESSAGE, null, answerOptions,
+                                        answerOptions[0]);
+                            }
+                            if ((deleteAll) || (dialogResult == JOptionPane.YES_OPTION)) {
+                                deleteFile(source, db_path, path);
+                            } else if (dialogResult == JOptionPane.NO_OPTION) {
+                                deleteAll = true;
+                                deleteFile(source, db_path, path);
+                            }
+
                         } else {
-                            tree.collapsePath(path);
-                            tree.expandPath(path);
+                            JOptionPane.showMessageDialog(null, "You cannot delete non-empty directories!",
+                                    "BaseX Delete Warning", JOptionPane.PLAIN_MESSAGE);
                         }
-                    }
-                    if (treeModel.getChildCount(path.getLastPathComponent()) == 0) {
-
-                        int dialogResult = JOptionPane.YES_OPTION;
-                        if (!deleteAll) {
-                            Object[] answerOptions = {"Yes" , "All" , "No"};
-                            dialogResult = JOptionPane.showOptionDialog(null, "Do you really want to delete the file\n" +
-                                            TreeUtils.urlStringFromTreePath(path) + "?", "Delete Resource(s)",
-                                    JOptionPane.YES_NO_CANCEL_OPTION, JOptionPane.QUESTION_MESSAGE, null, answerOptions,
-                                    answerOptions[0]);
-                        }
-                        if ((deleteAll) || (dialogResult == JOptionPane.YES_OPTION)) {
-                            deleteFile(source, db_path, path);
-                        } else if (dialogResult == JOptionPane.NO_OPTION) {
-                            deleteAll = true;
-                            deleteFile(source, db_path, path);
-                        }
-
                     } else {
-                        JOptionPane.showMessageDialog(null, "You cannot delete non-empty directories!",
+                        JOptionPane.showMessageDialog(null, "You cannot delete databases!",
                                 "BaseX Delete Warning", JOptionPane.PLAIN_MESSAGE);
                     }
-                } else {
-                    JOptionPane.showMessageDialog(null, "You cannot delete databases!",
-                            "BaseX Delete Warning", JOptionPane.PLAIN_MESSAGE);
                 }
             }
         }
